+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.19.1
 - Fix logging in `on_ready`
 - Log an error if some lavalink files are not found
+- Fix lavalink_request
 
 ## 0.19.0
 - Ether disconnect when using the command `stop`